requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# NLP
nltk>=3.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web interface
streamlit>=1.28.0

# CLI
click>=8.1.0

# Configuration
pyyaml>=6.0

# Jupyter (for exploration)
jupyter>=1.0.0
ipykernel>=6.25.0
